utils: add a gstreamer helper application for interconnection with pipewire
Unfortunately, the bluez-alsa PCM plugin does not work correctly
when it is used through pipewire (or gstreamer, or anywhere really...).

Thanfully, the bluez-alsa PCM plugin is only a simple client that
reads/writes on a file descriptor that was opened by bluealsa.
This allows us to use bluealsa without the PCM plugin, just like it
is done in the aplay.c util.

This one uses GStreamer to implement the plumbing between pipewire
and the file descriptor. On the reading side we are also doing some
tricks to ensure a smooth stream, which is not the case for the
stream that is coming out of bluealsa.

This helper is implemented as a patch to bluez-alsa so that it can
use its internal private API. In the future this needs some re-thinking.

Upstream-Status: Inappropriate